Split YakPlayer into YakPlayer and VerwaltungsYak
