--- src/main/java/de/YakWeide/YakPlayerUtility/PropertyList.java
@@ -1,4 +1,0 @@
-package de.YakWeide.YakPlayerUtility;
-
-public class PropertyList {
-}
