--- src/main/java/de/YakWeide/YakPlayerUtility/Yak.java
@@ -1,4 +1,0 @@
-package de.YakWeide.YakPlayerUtility;
-
-public class Yak implements YakPlayer{
-}
