+++ src/main/java/de/YakWeide/YakPlayerUtility/YakAttribute.java
@@ -1,0 +1,6 @@
+package de.YakWeide.YakPlayerUtility;
+
+public class YakAttribute {
+    private static int maxOfThisProperty;
+
+}
